Move pagination links into separate component
This allows us to remove the "use client" directive
from [[...id]]/page.tsx
